+++ trabalhoSD/src/main/java/cliente/ClientApp.java
@@ -1,0 +1,13 @@
+package cliente;
+
+public class ClientApp {
+	
+	public static void main(String args[]) {
+		try {
+			Thread cliente = new Thread(new Client());
+			cliente.start();
+			cliente.join();
+		} catch (Exception e) {
+		}
+	}
+}
