--- trabalhoSD/src/main/java/facomSD/facom/com/App.java
@@ -1,13 +1,0 @@
-package facomSD.facom.com;
-
-/**
- * Hello world!
- *
- */
-public class App 
-{
-    public static void main( String[] args )
-    {
-        System.out.println( "Hello World!" );
-    }
-}
