+++ trabalhoSD/src/main/java/servidor/ServerApp.java
@@ -1,0 +1,9 @@
+package servidor;
+
+public class ServerApp {
+
+	public static void main(String[] args) {
+		Server server = new Server();
+		server.iniciar();
+	}
+}
